Sort By function under Window Menu
Added sort by function under window menu for issue #10393

+++ PowerEditor/src/WinControls/WindowsDlg/WindowsDlg.cpp
@@ -948,7 +948,7 @@ void WindowsDlg::doCount()
 
 void WindowsDlg::doSort()
 {
-	UINT count = (_pTab != NULL) ? _pTab->nbItem() : 0;
+	size_t count = (_pTab != NULL) ? _pTab->nbItem() : 0;
 	auto currrentTabIndex = _pTab->getCurrentTabIndex();
 	NMWINDLG nmdlg;
 	nmdlg.type = WDT_SORT;
